refactor: modularize server and audio logic
