+++ src/index.js
@@ -4,13 +4,12 @@
 const fs = require("fs");
 const path = require("path");
 const storage = require("node-persist");
-const { spawn } = require("child_process");
-const fetch = require("node-fetch");
-const express = require("express");
 
 const packageJson = require("../package.json");
 const options = require("./utils/options.js");
 
+const { registerServer } = require("./server");
+const { registerAudio } = require("./audio");
 // HomeKit error
 const HK_NOT_ALLOWED_IN_CURRENT_STATE = -70412;
 
@@ -21,7 +20,6 @@ const originTypes = {
   EXTERNAL: 4,
 };
 
-const app = express();
 let Service, Characteristic, storagePath;
 
 module.exports = function (homebridge) {
@@ -29,6 +27,8 @@ module.exports = function (homebridge) {
   Characteristic = homebridge.hap.Characteristic;
   storagePath = homebridge.user.storagePath();
 
+  registerServer(SecuritySystem, Characteristic, originTypes);
+  registerAudio(SecuritySystem, Characteristic, originTypes);
   homebridge.registerAccessory(
     "homebridge-securitysystem",
     "security-system",
@@ -1395,528 +1395,6 @@ SecuritySystem.prototype.updateTripSwitch = function (
   return true;
 };
 
-// Server
-SecuritySystem.prototype.isAuthenticated = function (req, res) {
-  // Check if authentication is disabled
-  if (options.serverCode === null) {
-    return null;
-  }
-
-  let code = req.query.code;
-
-  // Check if code sent
-  if (code === undefined) {
-    this.sendCodeRequiredError(res);
-    return false;
-  }
-
-  // Check brute force
-  if (this.invalidCodeCount >= 5) {
-    req.blocked = true;
-    this.sendCodeInvalidError(req, res);
-    return false;
-  }
-
-  const userCode = parseInt(req.query.code);
-
-  if (userCode !== options.serverCode) {
-    this.invalidCodeCount++;
-    this.sendCodeInvalidError(req, res);
-    return false;
-  }
-
-  // Reset
-  this.invalidCodeCount = 0;
-
-  return true;
-};
-
-SecuritySystem.prototype.getDelayParameter = function (req) {
-  return req.query.delay === "true" ? true : false;
-};
-
-SecuritySystem.prototype.sendCodeRequiredError = function (res) {
-  this.log.info("Code required (Server)");
-
-  const response = {
-    error: true,
-    message: "Code required",
-    hint: "Add the 'code' URL parameter with your security code",
-  };
-
-  res.status(401).json(response);
-};
-
-SecuritySystem.prototype.sendCodeInvalidError = function (req, res) {
-  const response = { error: true };
-
-  if (req.blocked) {
-    this.log.info("Code blocked (Server)");
-    response.message = "Code blocked";
-  } else {
-    this.log.info("Code invalid (Server)");
-    response.message = "Code invalid";
-  }
-
-  res.status(403).json(response);
-};
-
-SecuritySystem.prototype.sendResultResponse = function (res, success) {
-  const response = {
-    error: success ? false : true,
-  };
-
-  res.json(response);
-};
-
-SecuritySystem.prototype.startServer = async function () {
-  app.get("/", (req, res) => {
-    res.redirect(
-      "https://github.com/MiguelRipoll23/homebridge-securitysystem/wiki/Server"
-    );
-  });
-
-  app.get("/status", (req, res) => {
-    if (this.isAuthenticated(req, res) === false) {
-      return;
-    }
-
-    const response = {
-      arming: this.isArming,
-      current_mode: this.state2Mode(this.currentState),
-      target_mode: this.state2Mode(this.targetState),
-      tripped: this.triggerTimeout !== null,
-    };
-
-    res.json(response);
-  });
-
-  app.get("/triggered", (req, res) => {
-    if (this.isAuthenticated(req, res) === false) {
-      return;
-    }
-
-    let result = true;
-
-    if (this.getDelayParameter(req)) {
-      // Delay
-      result = this.updateTripSwitch(true, originTypes.EXTERNAL, false, null);
-    } else {
-      const isCurrentStateDisarmed =
-        this.currentState ===
-        Characteristic.SecuritySystemCurrentState.DISARMED;
-
-      // Not armed
-      if (isCurrentStateDisarmed && options.overrideOff === false) {
-        this.sendResultResponse(res, false);
-        return;
-      }
-
-      this.setCurrentState(
-        Characteristic.SecuritySystemCurrentState.ALARM_TRIGGERED,
-        originTypes.EXTERNAL
-      );
-    }
-
-    this.sendResultResponse(res, result);
-  });
-
-  app.get("/home", (req, res) => {
-    if (this.isAuthenticated(req, res) === false) {
-      return;
-    }
-
-    const state = Characteristic.SecuritySystemTargetState.STAY_ARM;
-    const delay = this.getDelayParameter(req);
-    const result = this.updateTargetState(
-      state,
-      originTypes.EXTERNAL,
-      delay,
-      null
-    );
-
-    this.sendResultResponse(res, result);
-  });
-
-  app.get("/away", (req, res) => {
-    if (this.isAuthenticated(req, res) === false) {
-      return;
-    }
-
-    const state = Characteristic.SecuritySystemTargetState.AWAY_ARM;
-    const delay = this.getDelayParameter(req);
-    const result = this.updateTargetState(
-      state,
-      originTypes.EXTERNAL,
-      delay,
-      null
-    );
-
-    this.sendResultResponse(res, result);
-  });
-
-  app.get("/night", (req, res) => {
-    if (this.isAuthenticated(req, res) === false) {
-      return;
-    }
-
-    const state = Characteristic.SecuritySystemTargetState.NIGHT_ARM;
-    const delay = this.getDelayParameter(req);
-    const result = this.updateTargetState(
-      state,
-      originTypes.EXTERNAL,
-      delay,
-      null
-    );
-
-    this.sendResultResponse(res, result);
-  });
-
-  app.get("/off", (req, res) => {
-    if (this.isAuthenticated(req, res) === false) {
-      return;
-    }
-
-    const state = Characteristic.SecuritySystemTargetState.DISARM;
-    const delay = this.getDelayParameter(req);
-    const result = this.updateTargetState(
-      state,
-      originTypes.EXTERNAL,
-      delay,
-      null
-    );
-
-    this.sendResultResponse(res, result);
-  });
-
-  app.get("/arming-lock/:mode/:value", (req, res) => {
-    if (this.isAuthenticated(req, res) === false) {
-      return;
-    }
-
-    const mode = req.params["mode"].toLowerCase();
-    const value = req.params["value"].includes("on");
-    const result = this.updateArmingLock(mode, value);
-
-    this.sendResultResponse(res, result);
-  });
-
-  // Listener
-  const server = app.listen(options.serverPort, (error) => {
-    if (error) {
-      this.log.error("Error while starting server.");
-      this.log.error(error);
-      return;
-    }
-
-    this.log.info(`Server (${options.serverPort})`);
-  });
-
-  server.on("error", (error) => {
-    this.log.error("Error while starting server.");
-    this.log.error(error);
-  });
-};
-
-// Audio
-SecuritySystem.prototype.playAudio = async function (type, state) {
-  // Check option
-  if (options.audio === false) {
-    return;
-  }
-
-  const mode = this.state2Mode(state);
-
-  // Close previous player
-  this.stopAudio();
-
-  // Ignore 'Current Off' event
-  if (mode === "off") {
-    if (type === "target") {
-      return;
-    }
-  }
-
-  // Check audio switch except for triggered
-  const audioSwitchOnCharacteristic = this.audioSwitchService.getCharacteristic(
-    Characteristic.On
-  );
-  const isAudioDisabledBySwitch = audioSwitchOnCharacteristic.value === false;
-
-  if (mode !== "triggered" && isAudioDisabledBySwitch) {
-    return;
-  }
-
-  // Directory
-  let directory = `${__dirname}/../sounds`;
-
-  if (options.isValueSet(options.audioPath)) {
-    directory = options.audioPath;
-
-    if (directory[directory.length] === "/") {
-      directory = directory.substring(0, directory.length - 1);
-    }
-  }
-
-  // Check if file exists
-  const filename = `${type}-${mode}.mp3`;
-  const filePath = `${directory}/${options.audioLanguage}/${filename}`;
-
-  try {
-    await fs.promises.access(filePath);
-  } catch (error) {
-    this.log.debug(`Sound file not found (${filePath})`);
-    return;
-  }
-
-  // Arguments
-  let commandArguments = ["-loglevel", "error", "-nodisp", "-i", `${filePath}`];
-
-  if (mode === "triggered") {
-    commandArguments.push("-loop");
-    commandArguments.push("-1");
-  } else if (
-    (mode === "home" || mode === "night" || mode === "away") &&
-    type === "target" &&
-    options.audioArmingLooped
-  ) {
-    commandArguments.push("-loop");
-    commandArguments.push("-1");
-  } else if (mode === "warning" && options.audioAlertLooped) {
-    commandArguments.push("-loop");
-    commandArguments.push("-1");
-  } else {
-    commandArguments.push("-autoexit");
-  }
-
-  if (options.isValueSet(options.audioVolume)) {
-    commandArguments.push("-volume");
-    commandArguments.push(options.audioVolume);
-  }
-
-  // Process
-  const environmentVariables = [process.env];
-
-  options.audioExtraVariables.forEach((variable) => {
-    const key = variable.key;
-    const value = variable.value;
-    environmentVariables[key] = value;
-  });
-
-  this.log.debug("Environment Variables (Audio)", environmentVariables);
-
-  const ffplayEnv = {
-    ...process.env,
-    ...environmentVariables,
-  };
-
-  this.audioProcess = spawn("ffplay", commandArguments, { env: ffplayEnv });
-  this.log.debug(`ffplay ${commandArguments.join(" ")}`);
-
-  this.audioProcess.on("error", (data) => {
-    // Check if command is missing
-    if (data !== null && data.toString().indexOf("ENOENT") > -1) {
-      this.log.error("Unable to play sound, ffmpeg is not installed.");
-      return;
-    }
-
-    this.log.error(`Unable to play sound.\n${data}`);
-  });
-
-  this.audioProcess.on("close", function () {
-    this.audioProcess = null;
-  });
-};
-
-SecuritySystem.prototype.stopAudio = function () {
-  if (this.audioProcess !== null) {
-    this.audioProcess.kill();
-  }
-};
-
-SecuritySystem.prototype.setupAudio = async function () {
-  try {
-    await fs.promises.access(`${options.audioPath}/${options.audioLanguage}`);
-  } catch (error) {
-    await fs.promises.mkdir(`${options.audioPath}/${options.audioLanguage}`);
-    await fs.promises.copyFile(
-      `${__dirname}/sounds/README`,
-      `${options.audioPath}/README`
-    );
-    await fs.promises.copyFile(
-      `${__dirname}/sounds/README`,
-      `${options.audioPath}/README.txt`
-    );
-
-    this.log.warn("Check audio path directory for instructions.");
-  }
-};
-
-// Command
-SecuritySystem.prototype.executeCommand = function (type, state, origin) {
-  // Check proxy mode
-  if (options.proxyMode && origin === originTypes.EXTERNAL) {
-    this.log.debug("Command bypassed as proxy mode is enabled.");
-    return;
-  }
-
-  let command = null;
-
-  switch (state) {
-    case Characteristic.SecuritySystemCurrentState.ALARM_TRIGGERED:
-      command = options.commandCurrentTriggered;
-      break;
-
-    case Characteristic.SecuritySystemCurrentState.STAY_ARM:
-      if (type === "current") {
-        command = options.commandCurrentHome;
-        break;
-      }
-
-      command = options.commandTargetHome;
-      break;
-
-    case Characteristic.SecuritySystemCurrentState.AWAY_ARM:
-      if (type === "current") {
-        command = options.commandCurrentAway;
-        break;
-      }
-
-      command = options.commandTargetAway;
-      break;
-
-    case Characteristic.SecuritySystemCurrentState.NIGHT_ARM:
-      if (type === "current") {
-        command = options.commandCurrentNight;
-        break;
-      }
-
-      command = options.commandTargetNight;
-      break;
-
-    case Characteristic.SecuritySystemCurrentState.DISARMED:
-      if (type === "current") {
-        command = options.commandCurrentOff;
-        break;
-      }
-
-      command = options.commandTargetOff;
-      break;
-
-    case "warning":
-      command = options.commandCurrentWarning;
-      break;
-
-    default:
-      this.log.error(`Unknown command ${type} state (${state})`);
-  }
-
-  if (command === undefined || command === null) {
-    this.log.debug(`Command option for ${type} mode is not set.`);
-    return;
-  }
-
-  // Parameters
-  command = command.replace(
-    "${currentMode}",
-    this.state2Mode(this.currentState)
-  );
-
-  const process = spawn(command, { shell: true });
-
-  process.stderr.on("data", (data) => {
-    this.log.error(`Command failed (${command})\n${data}`);
-  });
-
-  process.stdout.on("data", (data) => {
-    this.log.info(`Command output: ${data}`);
-  });
-};
-
-// Webhooks
-SecuritySystem.prototype.sendWebhookEvent = function (type, state, origin) {
-  // Check webhook host
-  if (options.isValueSet(options.webhookUrl) === false) {
-    this.log.debug("Webhook base URL option is not set.");
-    return;
-  }
-
-  // Check proxy mode
-  if (options.proxyMode && origin === originTypes.EXTERNAL) {
-    this.log.debug("Webhook bypassed as proxy mode is enabled.");
-    return;
-  }
-
-  let path = null;
-
-  switch (state) {
-    case Characteristic.SecuritySystemCurrentState.ALARM_TRIGGERED:
-      path = options.webhookCurrentTriggered;
-      break;
-
-    case Characteristic.SecuritySystemCurrentState.STAY_ARM:
-      if (type === "current") {
-        path = options.webhookCurrentHome;
-        break;
-      }
-
-      path = options.webhookTargetHome;
-      break;
-
-    case Characteristic.SecuritySystemCurrentState.AWAY_ARM:
-      if (type === "current") {
-        path = options.webhookCurrentAway;
-        break;
-      }
-
-      path = options.webhookTargetAway;
-      break;
-
-    case Characteristic.SecuritySystemCurrentState.NIGHT_ARM:
-      if (type === "current") {
-        path = options.webhookCurrentNight;
-        break;
-      }
-
-      path = options.webhookTargetNight;
-      break;
-
-    case Characteristic.SecuritySystemCurrentState.DISARMED:
-      if (type === "current") {
-        path = options.webhookCurrentOff;
-        break;
-      }
-
-      path = options.webhookTargetOff;
-      break;
-
-    case "warning":
-      path = options.webhookCurrentWarning;
-      break;
-
-    default:
-      this.log.error(`Unknown webhook ${type} state (${state})`);
-      return;
-  }
-
-  if (path === undefined || path === null) {
-    this.log.debug(`Webhook option for ${type} mode is not set.`);
-    return;
-  }
-
-  // Parameters
-  path = path.replace("${currentMode}", this.state2Mode(this.currentState));
-
-  // Send GET request to server
-  fetch(options.webhookUrl + path)
-    .then((response) => {
-      if (response.ok === false) {
-        throw new Error(`Status code (${response.status})`);
-      }
-
-      this.log.info("Webhook event (Sent)");
-    })
-    .catch((error) => {
       this.log.error(`Request to webhook failed. (${path})`);
       this.log.error(error);
     });
